requirements.txt
gym[atari]
gym==0.10.9
matplotlib==2.0.2
numpy==1.14.2
six==1.11.0
tensorflow-gpu==1.7.1
scikit-learn==0.20.2
wandb
